check out a tag properly

+++ scripts/root/new/install_w_cmake.sh
@@ -13,26 +13,33 @@ cd $wdir
 git clone http://root.cern.ch/git/root.git
 
 #version=v6-05-02
-version=v5-34-32
+#version=v5-34-32
+version=v5-34-34
 tdir=$HOME/software/root/$version
 
 cd root
 git tag -l
-git checkout -b $version $version
+#git checkout -b $version 
+#$version
 
-cd $wdir
+git checkout tags/$version
 
-dname=`date +"%Y-%m-%d"`
-bdir="build_$version_$dname"
-rm -rf $bdir
-mkdir $bdir
-cd $bdir
+if [ "$1" = "build" ]; then
 
-cmake $wdir/root
-cmake --build .
-cmake -DCMAKE_INSTALL_PREFIX=$tdir -P cmake_install.cmake
-
-ln -s $tdir/bin/thisroot.sh "/usr/local/bin/root-$version"
+	cd $wdir
+	
+	dname=`date +"%Y-%m-%d"`
+	bdir="build_$version_$dname"
+	rm -rf $bdir
+	mkdir $bdir
+	cd $bdir
+	
+	cmake $wdir/root
+	cmake --build .
+	cmake -DCMAKE_INSTALL_PREFIX=$tdir -P cmake_install.cmake
+	
+	ln -s $tdir/bin/thisroot.sh "/usr/local/bin/root-$version"
+fi
 
 cd $savedir
